Do req/res serialization in validate function
Remove req/res serialization from ob-proxy.js file.
